requirements.txt
hydra-core
pypdf
tiktoken
pytesseract
pdf2image
Pillow